requirements.txt
masonite>=2.1.0,<2.2.0
pytest
codecov
PyJWT
psycopg2-binary
python-slugify
masonite-api
pytest-cov 